backend/requirements.txt
openai==1.57.0
langchain==0.3.7
pillow==11.0.0
pydantic==2.9.2
tiktoken==0.8.0
poppler-utils==0.1.0
pytesseract==0.3.13
nltk==3.9.1
tesseract==0.1.3
python-dotenv==1.0.1
ollama==0.3.2
langchain-openai==0.2.11
langchain-community==0.3.5
langchain-ollama==0.2.0
mysql-connector-python
llama-stack
docx2pdf==0.1.8
PyPDF2==3.0.1
extract-msg==0.52.0
SQLAlchemy==2.0.3
mysql-connector-python==9.1.0
fastapi==0.115.4
uvicorn==0.32.0
langserve==0.3.1
sse-starlette==2.2.1
PyMuPDF==1.25.2
PyMySQL==1.1.1
pdfminer
pdfminer.six
python-multipart
langchain-google-genai
google-generativeai
celery
pika